fraud-backend-2/requirements.txt
# Enhanced URL Fraud Detection System Requirements

# Core Flask components
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# HTTP and Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# Domain analysis
python-whois>=0.7.0

# Visual analysis
selenium>=4.0.0
opencv-python>=4.5.0
pillow>=9.0.0
imagehash>=4.2.0

# Machine learning
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.3.0
joblib>=1.1.0

# Browser automation
chromedriver-autoinstaller>=0.6.0

# Development and testing
pytest>=7.0.0
pytest-flask>=1.2.0

# Optional: Environment management
python-dotenv>=0.19.0
